refactor: strategy test use cases
