+++ packages/rs-drive-abci/tests/strategy_tests/test_cases/withdrawal_tests.rs
@@ -1,12 +1,12 @@
 mod tests {
-    use crate::asset_unlock_index;
     use crate::execution::{continue_chain_for_strategy, run_chain_for_strategy, GENESIS_TIME_MS};
     use crate::strategy::{
         ChainExecutionOutcome, ChainExecutionParameters, NetworkStrategy, StrategyRandomness,
     };
     use dpp::dashcore::bls_sig_utils::BLSSignature;
     use dpp::dashcore::hashes::Hash;
-    use dpp::dashcore::{BlockHash, ChainLock, Txid};
+    use dpp::dashcore::transaction::TransactionPayload::AssetUnlockPayloadType;
+    use dpp::dashcore::{BlockHash, ChainLock, Transaction, Txid};
     use dpp::dashcore_rpc::dashcore_rpc_json::{AssetUnlockStatus, AssetUnlockStatusResult};
     use dpp::data_contracts::withdrawals_contract;
     use dpp::identity::{KeyType, Purpose, SecurityLevel};
@@ -31,6 +31,13 @@ mod tests {
     use strategy_tests::frequency::Frequency;
     use strategy_tests::operations::{Operation, OperationType};
     use strategy_tests::{IdentityInsertInfo, StartIdentities, Strategy};
+
+    fn asset_unlock_index(tx: &Transaction) -> u64 {
+        let Some(AssetUnlockPayloadType(ref payload)) = tx.special_transaction_payload else {
+            panic!("expected to get AssetUnlockPayloadType");
+        };
+        payload.base.index
+    }
 
     struct CoreState {
         asset_unlock_statuses: BTreeMap<WithdrawalTransactionIndex, AssetUnlockStatusResult>,
